anh96

+++ ApiProject/Mgm.Application/PositionHistory/Dtos/PHistoryDeteail.cs
@@ -4,8 +4,8 @@ namespace Mgm.PositionHistory.Dtos
 {
     public class PHistoryDeteail
     {
-        public int Lng { get; set; }
-        public int Lat { get; set; }
+        public double Lng { get; set; }
+        public double Lat { get; set; }
         public string Address { get; set; }
         public int IsWarning { get; set; }
         public DateTime VerifyDate { get; set; }
