+++ ApiProject/Mgm.Core/Covid19/PositionHistoryDetail/PositionHistoryDetail.cs
@@ -8,8 +8,8 @@ namespace Mgm.Covid19.PositionHistoryDetail
     public class PositionsHistoryDetail : Entity
     {
         public int TimeOutId { get; set; }
-        public int Lng { get; set; }
-        public int Lat { get; set; }
+        public double Lng { get; set; }
+        public double Lat { get; set; }
         public string Address { get; set; }
         public int IsWarning { get; set; }
         public DateTime VerifyDate { get; set; }
